requirements.txt
opencv-python==4.5.1.48
pytorch==1.10.1